backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0

# Anthropic
anthropic==0.34.2

# Redis
redis==5.0.8

# MCP
mcp==1.0.0

# HTTP clients (for MCP servers)
httpx==0.27.2

# Logging
structlog==24.4.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-mock==3.14.0
httpx==0.27.2
